requirements.txt
Flask==2.0.1
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
mysqlclient==2.1.1
python-dotenv==0.19.0
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.1
Flask-Supabase==0.1.0
